refactor: inner class name public class로 변환
